reverted *_permits options back to `usize`

+++ rust/crates/rust_mdex_dl/src/config.rs
@@ -46,8 +46,9 @@ pub struct Client {
 #[allow(unused)]
 #[derive(Deserialize, Debug, Clone)]
 pub struct Concurrency {
-    pub image_permits: u32,
-    pub chapter_permits: u32,
+    // semaphores take `usize`, so don't use `u32` here
+    pub image_permits: usize,
+    pub chapter_permits: usize,
 }
 
 #[allow(unused)]
